Rename module from watchlist to following
