+++ app-features/following/build.gradle.kts
