--- app-features/watchlist/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.thomaskioko.tvmaniac.watchlist" />
