+++ app/src/main/kotlin/com/thomaskioko/tvmaniac/injection/NavigationModule.kt
@@ -2,11 +2,11 @@ package com.thomaskioko.tvmaniac.injection
 
 import com.thomaskioko.showdetails.ShowDetailNavigationFactory
 import com.thomaskioko.tvmaniac.DiscoverNavigationFactory
+import com.thomaskioko.tvmaniac.following.FollowingNavigationFactory
 import com.thomaskioko.tvmaniac.navigation.ComposeNavigationFactory
 import com.thomaskioko.tvmaniac.search.SearchNavigationFactory
 import com.thomaskioko.tvmaniac.settings.SettingsNavigationFactory
 import com.thomaskioko.tvmaniac.show_grid.ShowsGridNavigationFactory
-import com.thomaskioko.tvmaniac.watchlist.WatchlistNavigationFactory
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -31,7 +31,7 @@ internal interface NavigationModule {
     @Singleton
     @Binds
     @IntoSet
-    fun bindWatchlistNavigation(factory: WatchlistNavigationFactory): ComposeNavigationFactory
+    fun bindWatchlistNavigation(factory: FollowingNavigationFactory): ComposeNavigationFactory
 
     @Singleton
     @Binds
